feat: allow handlers to provide arbitrary format for the error responses (#15)
* feat: allow handlers to provide arbitrary format for the error responses

Before clients of the library were limited by the marshaling implementation of HTTPError which did not allow
to change the format of error responses. This is required to support migrattion of existing APIs for which clients
need to stay backward compatible.

* fix: install golangci-linter with the recommended approach

The previous one with `go get` is no more supported and not recomended esp. for older go versions.

* fix: review comments

rename interface to be public to include it in godocs, fix line wrapping.